Update Controller (6)

+++ Unibean.Service/Services/StoreService.cs
@@ -4,10 +4,10 @@ using Unibean.Repository.Entities;
 using Unibean.Repository.Paging;
 using Unibean.Repository.Repositories.Interfaces;
 using Unibean.Service.Models.Activities;
+using Unibean.Service.Models.CampaignDetails;
 using Unibean.Service.Models.Exceptions;
 using Unibean.Service.Models.Stores;
 using Unibean.Service.Models.Transactions;
-using Unibean.Service.Models.Vouchers;
 using Unibean.Service.Services.Interfaces;
 using Unibean.Service.Utilities.FireBase;
 using BCryptNet = BCrypt.Net.BCrypt;
@@ -32,7 +32,7 @@ public class StoreService : IStoreService
 
     private readonly IAccountRepository accountRepository;
 
-    private readonly IVoucherService voucherService;
+    private readonly ICampaignDetailService campaignDetailService;
 
     private readonly IActivityService activityService;
 
@@ -45,7 +45,7 @@ public class StoreService : IStoreService
     public StoreService(IStoreRepository storeRepository,
         IFireBaseService fireBaseService,
         IAccountRepository accountRepository,
-        IVoucherService voucherService,
+        ICampaignDetailService campaignDetailService,
         IActivityService activityService,
         IBonusService bonusService,
         IVoucherItemRepository voucherItemRepository,
@@ -112,7 +112,7 @@ public class StoreService : IStoreService
         this.storeRepository = storeRepository;
         this.fireBaseService = fireBaseService;
         this.accountRepository = accountRepository;
-        this.voucherService = voucherService;
+        this.campaignDetailService = campaignDetailService;
         this.activityService = activityService;
         this.bonusService = bonusService;
         this.voucherItemRepository = voucherItemRepository;
@@ -142,34 +142,39 @@ public class StoreService : IStoreService
         (string id, string voucherItemId, CreateUseActivityModel creation)
     {
         var item = voucherItemRepository.GetById(voucherItemId);
-        if (item.CampaignDetail.Campaign.CampaignStores.Any(c => c.StoreId.Equals(id)))
+        if (new[] { CampaignState.Active, CampaignState.Inactive }.Contains
+            (item.CampaignDetail.Campaign.CampaignActivities.LastOrDefault().State.Value))
         {
-            if ((bool)item.IsBought && item.Activities.FirstOrDefault() != null)
+            if (item.CampaignDetail.Campaign.CampaignStores.Any(c => c.StoreId.Equals(id)))
             {
-                if (!(bool)item.IsUsed)
+                if ((bool)item.IsBought && item.Activities.FirstOrDefault() != null)
                 {
-                    var stu = studentRepository.GetById
-                        (item.Activities.FirstOrDefault().StudentId);
-                    if (stu != null && stu.State.Equals(StudentState.Active))
+                    if (!(bool)item.IsUsed)
                     {
-                        CreateActivityModel create = mapper.Map<CreateActivityModel>(creation);
-                        create.StudentId = item.Activities.FirstOrDefault().StudentId;
-                        create.VoucherItemId = voucherItemId;
-                        create.StoreId = id;
-                        activityService.Add(create);
-                        return true;
+                        var stu = studentRepository.GetById
+                            (item.Activities.FirstOrDefault().StudentId);
+                        if (stu != null && stu.State.Equals(StudentState.Active))
+                        {
+                            CreateActivityModel create = mapper.Map<CreateActivityModel>(creation);
+                            create.StudentId = item.Activities.FirstOrDefault().StudentId;
+                            create.VoucherItemId = voucherItemId;
+                            create.StoreId = id;
+                            return activityService.Add(create) != null;
+                        }
+                        throw new InvalidParameterException
+                            ("Sinh viên không hợp lệ");
                     }
                     throw new InvalidParameterException
-                        ("Sinh viên không hợp lệ");
+                        ("Mã khuyến mãi đã được sử dụng");
                 }
                 throw new InvalidParameterException
-                    ("Mã khuyến mãi đã được sử dụng");
+                    ("Mã khuyến mãi chưa được thanh toán");
             }
             throw new InvalidParameterException
-                ("Mã khuyến mãi chưa được thanh toán");
+                ("Mã khuyến mãi không được áp dụng cho cửa hàng này");
         }
         throw new InvalidParameterException
-            ("Mã khuyến mãi không được áp dụng cho cửa hàng này");
+            ("Chiến dịch không hợp lệ");
     }
 
     public void Delete(string id)
@@ -177,15 +182,23 @@ public class StoreService : IStoreService
         Store entity = storeRepository.GetById(id);
         if (entity != null)
         {
-            // Avatar
-            if (entity.Account.Avatar != null && entity.Account.Avatar.Length > 0)
+            if (entity.CampaignStores.All(
+                s => s.Campaign.CampaignActivities.LastOrDefault().State.Equals(CampaignState.Closed)))
             {
-                // Remove image
-                fireBaseService.RemoveFileAsync(entity.Account.FileName, ACCOUNT_FOLDER_NAME);
-            }
+                // Avatar
+                if (entity.Account.Avatar != null && entity.Account.Avatar.Length > 0)
+                {
+                    // Remove image
+                    fireBaseService.RemoveFileAsync(entity.Account.FileName, ACCOUNT_FOLDER_NAME);
+                }
 
-            storeRepository.Delete(id);
-            accountRepository.Delete(entity.Account.Id);
+                storeRepository.Delete(id);
+                accountRepository.Delete(entity.Account.Id);
+            }
+            else
+            {
+                throw new InvalidParameterException("Xóa thất bại do tồn tại chiến dịch hoạt động ở cửa hàng");
+            }
         }
         else
         {
@@ -255,15 +268,15 @@ public class StoreService : IStoreService
         throw new InvalidParameterException("Không tìm thấy cửa hàng");
     }
 
-    public PagedResultModel<VoucherModel> GetVoucherListByStoreId
+    public PagedResultModel<CampaignDetailModel> GetCampaignDetailByStoreId
         (string id, List<string> campaignIds, List<string> typeIds, bool? state,
         string propertySort, bool isAsc, string search, int page, int limit)
     {
         Store entity = storeRepository.GetById(id);
         if (entity != null)
         {
-            return voucherService.GetAllByStore
-                (new() { id }, campaignIds, typeIds, state, propertySort, isAsc, search, page, limit);
+            return campaignDetailService.GetAllByStore
+                (id, campaignIds, typeIds, state, propertySort, isAsc, search, page, limit);
         }
         throw new InvalidParameterException("Không tìm thấy cửa hàng");
     }
@@ -288,6 +301,16 @@ public class StoreService : IStoreService
             }
 
             return mapper.Map<StoreExtraModel>(storeRepository.Update(entity));
+        }
+        throw new InvalidParameterException("Không tìm thấy cửa hàng");
+    }
+
+    public CampaignDetailExtraModel GetCampaignDetailById(string id, string detailId)
+    {
+        Store entity = storeRepository.GetById(id);
+        if (entity != null)
+        {
+            return campaignDetailService.GetById(detailId);
         }
         throw new InvalidParameterException("Không tìm thấy cửa hàng");
     }
